productapi/requirements.txt
asgiref==3.3.4
astroid==2.4.2
Django==3.2.3
django-filter==2.4.0
djangorestframework==3.12.4
image==1.5.33
isort==5.7.0
lazy-object-proxy==1.4.3
Markdown==3.3.4
mccabe==0.6.1
Pillow==8.2.0
pylint==2.6.0
python-dateutil==2.8.1
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
wrapt==1.12.1
